change OccupancyCalculationRequest to take map as input, use BigDecimal instead of Integer, fix bug for single provided only one category

+++ src/main/java/io/github/okraskat/room/optimizer/domain/CalculatedOccupancy.java
@@ -1,4 +1,6 @@
 package io.github.okraskat.room.optimizer.domain;
 
-public record CalculatedOccupancy(RoomCategory roomCategory, int occupiedRooms, int potentialIncome) {
+import java.math.BigDecimal;
+
+public record CalculatedOccupancy(RoomCategory roomCategory, int occupiedRooms, BigDecimal potentialIncome) {
 }
