+++ src/main/java/io/github/okraskat/room/optimizer/domain/OccupancyCalculator.java
@@ -2,20 +2,21 @@ package io.github.okraskat.room.optimizer.domain;
 
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 @Service
 class OccupancyCalculator implements OccupancyCalculatorApi {
 
     @Override
-    public List<CalculatedOccupancy> calculate(Map<RoomCategory, Integer> availableRoomsPerCategory, Collection<Integer> potentialPayments) {
-        TreeSet<Integer> sortedPayments = new TreeSet<>(potentialPayments);
+    public List<CalculatedOccupancy> calculate(Map<RoomCategory, Integer> availableRoomsPerCategory, Collection<BigDecimal> potentialPayments) {
+        TreeSet<BigDecimal> sortedPayments = new TreeSet<>(potentialPayments);
         List<RoomCategory> categoriesSortedByPriceDesc = Arrays.asList(RoomCategory.values());
         categoriesSortedByPriceDesc.sort(Comparator.comparing(RoomCategory::getLowestAvailablePrice).reversed());
         OccupancyCalculation occupancyCalculation = new OccupancyCalculation(availableRoomsPerCategory, categoriesSortedByPriceDesc);
 
         while (!sortedPayments.isEmpty()) {
-            int highestPayment = Objects.requireNonNull(sortedPayments.pollLast());
+            BigDecimal highestPayment = Objects.requireNonNull(sortedPayments.pollLast());
             categoriesSortedByPriceDesc.stream()
                     .filter(c -> c.isPaymentInCategoryRange(highestPayment))
                     .findFirst()
